requirements.txt
Flask
gunicorn
psycopg2-binary
streamlit
pandas